# Improve ProbabilisticMap performance for small value sets