Since `IndexOfAnyExcept` does a `value.Contains(c)` on every input character, it's faster if the value is at least 8 characters long.

We could go further with e.g. working with a `Vector128<ushort>` directly instead of calling into `Contains`, but I don't know how common short inputs will actually be in practice.

|            Method | Toolchain | Length |     Mean |    Error | Ratio |
|------------------ |---------- |------- |---------:|---------:|------:|
| IndexOfAnyExcept6 |      main |  10000 | 82.06 us | 0.288 us |  1.00 |
| IndexOfAnyExcept6 |        pr |  10000 | 31.00 us | 0.491 us |  0.38 |